+++ lib/stateclone/state_object.go
@@ -1,4 +1,4 @@
-package sebak
+package stateclone
 
 type DirtyStatus byte
